fix(compiler): fix float type compiler
Yet another difference between CEL and Firestore's rules language. Float is 'float' in Firestore,
and 'double' in CEL.

+++ src/compile/cel/types.ts
@@ -10,7 +10,7 @@ export enum CELGlobals {
   STRING = 'string',
   BOOLEAN = 'bool',
   INTEGER = 'int',
-  FLOAT = 'double',
+  FLOAT = 'float',
   LIST = 'list'
 }
